skr-without-sc-integration-dev - Part 2 (#14071)
* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* waitForApplicationCr

* clean up

* disable check for threshold

* enable check for threshold

* revert changes

* disable

* additional event checks

* removed parameters

* redudant value

* move function from migration test that will be deleted

* move function from migration test that will be deleted

* lint fix

* use dev client ID

* use single place for helper functions

* use single place for helper functions

* use single place for helper functions

* move to dedicated module of helper functions

* fix

* fix

* fix
