+++ tests/fast-integration/skr-test-without-sc/skr-test.js
@@ -7,7 +7,8 @@ const {
   ensureCommerceMockWithCompassTestFixture,
   checkFunctionResponse,
   sendLegacyEventAndCheckResponse,
-  deleteMockTestFixture,
+  deleteMockTestFixture, sendCloudEventStructuredModeAndCheckResponse, sendCloudEventBinaryModeAndCheckResponse,
+  checkInClusterEventDelivery,
 } = require('../test/fixtures/commerce-mock');
 const {
   ensureKymaAdminBindingExistsForUser,
@@ -17,7 +18,6 @@ const {
   AuditLogCreds,
   AuditLogClient,
   checkAuditLogs,
-  checkAuditEventsThreshold,
 } = require('../audit-log');
 const {keb, gardener, director} = require('./helpers');
 const {KCPWrapper, KCPConfig} = require('../kcp/client');
@@ -123,12 +123,24 @@ function commerceMockTest() {
       );
     });
 
+    it('in-cluster event should be delivered (structured and binary mode)', async function() {
+      await checkInClusterEventDelivery(this.options.testNS);
+    });
+
     it('function should be reachable through secured API Rule', async function() {
       await checkFunctionResponse(this.options.testNS);
     });
 
     it('order.created.v1 legacy event should trigger the lastorder function', async function() {
       await sendLegacyEventAndCheckResponse();
+    });
+
+    it('order.created.v1 cloud event in structured mode should trigger the lastorder function', async function() {
+      await sendCloudEventStructuredModeAndCheckResponse();
+    });
+
+    it('order.created.v1 cloud event in binary mode should trigger the lastorder function', async function() {
+      await sendCloudEventBinaryModeAndCheckResponse();
     });
 
     it('Deletes the resources that have been created', async function() {
@@ -140,12 +152,14 @@ function commerceMockTest() {
       const auditlogs = new AuditLogClient(AuditLogCreds.fromEnv());
 
       it('Check audit logs', async function() {
-        await checkAuditLogs(auditlogs, null, true);
+        await checkAuditLogs(auditlogs, null);
       });
 
-      it('Amount of audit events must not exceed a certain threshold', async function() {
-        await checkAuditEventsThreshold(4);
-      });
+      // TODO: Enable checkAuditEventsThreshold again when fix is ready by Andreas Thaler
+
+      // it('Amount of audit events must not exceed a certain threshold', async function() {
+      //   await checkAuditEventsThreshold(4);
+      // });
     }
   });
 }
